refactor(transit-display): extract TransitDisplay
Rename `TransitDisplay2` -> `TransitDisplay` (component + Props) and
move it out of `transit-displays-2.tsx` into its own
`transit-display.tsx`, mirroring the earlier `TransitDisplayEntry`
extraction.

Changes:

- New file `transit-display.tsx`: `TransitDisplay`, `TransitDisplayProps`,
  `StatsBadges` (private helper), and the TransitDisplay-only constants
  (`HeaderBadgeStyle`, `HEADER_STATS_ICONS_BY_SIZE`, plus a local copy
  of `BOARD_PANEL_BG` / `TITLE_TEXT_CLASS_BY_SIZE` /
  `TITLE_ICON_CLASS_BY_SIZE` that are still also referenced by the
  category-filter row in `transit-displays-2.tsx`).
- `transit-displays-2.tsx`: drop the moved component, Props, helper,
  and the TransitDisplay-only constants; import `TransitDisplay` from
  the new file; sheds the imports that only the moved code needed.
- `transit-display-per-route.tsx`: switch import to the new file and
  rename usages / TSDoc references.
- `transit-display-entry.stories.tsx`: update one TSDoc reference.
- Delete `transit-display-2.stories.tsx` (stories for the new
  `transit-display.tsx` can be added separately later).
